#66 refactor(cache): holidayStore를 repositories/caches/로 이관 + i18n 결합 제거
- holidayStore.ts 삭제 → repositories/caches/holidayCache.ts 신설
  - useHolidayStore → useHolidayCache 로 식별자 변경
  - @internal 주석 추가
  - i18n import 제거. fetchHolidays/refreshHolidays의 locale을 인자로 받도록 변경(optional, 기본값 'en')
  - setHolidaysForYear / clearYear / reset / setAvailableCountries 프리미티브 연산 추가
- 외부 사용처(MainCalendar, MainCalendarGrid, LeftSidebar, RightEventPanel, TopToolbar, HolidaySection) 모두 holidayCache로 import 갱신 + useHolidayCache 로 식별자 일괄 변경
- tests/stores/holidayStore.test.ts → holidayCache 경로 import 갱신
- tests/repositories/caches/holidayCache.test.ts 신설
